restructured header | added menu
