+++ shoppinglist-ui/src/app/app.component.ts
@@ -1,7 +1,4 @@
 import { ChangeDetectionStrategy, Component } from '@angular/core';
-import { Store } from '@ngrx/store';
-import { environment } from 'src/environments/environment';
-import { selectStatus } from './shopping-list/store/shopping-list.selectors';
 
 @Component({
   selector: 'app-root',
@@ -10,8 +7,5 @@ import { selectStatus } from './shopping-list/store/shopping-list.selectors';
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class AppComponent {
-  mock = environment.mock;
-  status$ = this.store.select(selectStatus);
-
-  constructor(private store: Store) {}
+  constructor() {}
 }
