--- shoppinglist-ui/src/app/header/header.component.html
@@ -1,11 +1,0 @@
-<h3>ShoppingList</h3>
-
-
-<ul nz-menu class="menu" nzTheme="dark" nzMode="horizontal">
-    <!-- <li nz-menu-item [routerLink]="['/']" >
-        <i type="message" nz-icon nzType="shopping-cart" class="logo"></i>
-    </li>
-    <li nz-menu-item [routerLink]="['shoppinglist']">
-    <i nz-icon nzType="shopping-cart"></i>
-    Einkaufsliste</li> -->
-</ul>
