--- shoppinglist-ui/src/app/header/header.component.less
@@ -1,23 +1,0 @@
-.logo {
-  text-align: center;
-  vertical-align: middle;
-  width: 120px;
-  height: 31px;
-  background: rgba(255, 255, 255, 0.2);
-  margin: 16px 24px 16px 0;
-  float: left;
-}
-
-h3 {
-  color: white;
-  vertical-align: middle;
-  margin: 0;
-}
-
-.menu {
-  width: 100%;
-}
-
-[nz-menu] {
-  line-height: 64px;
-}
