--- shoppinglist-ui/src/app/header/header.component.ts
@@ -1,12 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-header',
-  templateUrl: './header.component.html',
-  styleUrls: ['./header.component.less'],
-})
-export class HeaderComponent {
-  collapsed = true;
-
-  constructor() {}
-}
