+++ shoppinglist-ui/src/app/layout/header/header.component.less
@@ -1,0 +1,35 @@
+.header {
+  background-color: #001529;
+  height: 64px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  padding-left: 10px;
+}
+
+.name {
+  color: white;
+  margin: 0;
+}
+
+.menu-wrapper {
+  height: 100%;
+  width: 64px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.menu-icon {
+  color: white;
+  font-size: x-large;
+  margin: 0;
+}
+
+.info {
+  width: 64px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  height: 100%;
+}
